// src/app/shared/models/sitesetup.ts
export interface Site {

    activeStatus?: string;
    address?: string;
    city?: string;
    code?: string;
    companyCode?: string;
    contactPerson?: string;
    contactSalutation?: string;
    createdBy?: string;
    createdWhen?: string;
    email?: string;
    extCode1?: string;
    extCode1Remarks?: string;
    extCode2?: string;
    extCode2Remarks?: string;
    extension?: string;
    fax?: string;
    modifiedBy?: string;
    modifiedWhen?: string;
    name?: string;
    phone?: string;
    shortName?: string;
    supplyPointTag?: string;
}